requirements.txt
eventlet
Flask
Flask-Migrate
Flask-SQLAlchemy
psycopg2
python-dotenv
requests
Flask-Session
